requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
imageio==2.6.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.1.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numpy==1.17.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pybind11==2.4.3
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
pytesseract==0.3.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
PyWavelets==1.1.1
pyzmq==18.1.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
statsmodels==0.10.1
terminado==0.8.2
testpath==0.4.2
torch==1.3.0.post2
torchvision==0.4.1.post2
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.6.0
tqdm==4.38